Let `providers.systemProperty(propertyName)` have no value when `propertyName` has no value
