+++ subprojects/model-core/src/test/groovy/org/gradle/api/internal/provider/sources/EnvironmentVariableValueSourceTest.groovy
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.api.internal.provider.sources
+
+import org.gradle.api.internal.provider.ValueSourceBasedSpec
+
+class EnvironmentVariableValueSourceTest extends ValueSourceBasedSpec {
+
+    def "environment variable has no value when variable name has no value"() {
+        given:
+        def provider = createProviderOf(EnvironmentVariableValueSource) {
+            // give variable name no value
+        }
+
+        expect:
+        !provider.isPresent()
+    }
+}
